# cgifsave: palette change POC