Hey!

This is the POC of my approach to generating palettes for `cgifsave`.  How it works:

1. We compute a palette for each frame. The first frame's palette we use as a global one.
2. For each frame except the first we compare the palette computed in step 1 with the global palette and the local palette of the previous frame (if any).
3. If there's a local palette in the previous frame, and the diff between it and the current frame palette is small enough, we use the palette from the previous frame as local for the current frame.
4. Otherwise, if the diff between the global palette and the current frame palette is small enough, we don't use a local palette.
5. Otherwise, we use the palette computed in step 1 as a local palette for the current frame.

With this approach, we mitigate the difference between neighbor frame colors which helps transparency and frame optimizers do their jobs.

### Some numbers:

#### Master speeed:

```
$ /usr/bin/time -f %M:%e vips copy /images/animated_gif12.gif[n=-1] /images/animated_gif12_vips_out_master.gif[reoptimise=1]
40344:5.21

$ /usr/bin/time -f %M:%e vips copy /images/animated_gif13.gif[n=-1] /images/animated_gif13_vips_out_master.gif[reoptimise=1]
32552:5.33

$ /usr/bin/time -f %M:%e vips copy /images/animated_gif15.gif[n=-1] /images/animated_gif15_vips_out_master.gif[reoptimise=1]
27096:3.36

$ /usr/bin/time -f %M:%e vips copy /images/animated_gif16.gif[n=-1] /images/animated_gif16_vips_out_master.gif[reoptimise=1]
41744:2.18
```

#### New speeed:

```
$ /usr/bin/time -f %M:%e vips copy /images/animated_gif12.gif[n=-1] /images/animated_gif12_vips_out_new.gif[reoptimise=1]
40340:4.98

$ /usr/bin/time -f %M:%e vips copy /images/animated_gif13.gif[n=-1] /images/animated_gif13_vips_out_new.gif[reoptimise=1]
32972:5.11

$ /usr/bin/time -f %M:%e vips copy /images/animated_gif15.gif[n=-1] /images/animated_gif15_vips_out_new.gif[reoptimise=1]
26884:3.28

$ /usr/bin/time -f %M:%e vips copy /images/animated_gif16.gif[n=-1] /images/animated_gif16_vips_out_new.gif[reoptimise=1]
41592:2.06
```

#### Result sizes

Here I resized the images a bit to increase the color number.

```
$ ls -l /images/animated_gif*_vips_out_*
-rw-r--r-- 1 *** ***   687545 May 27 06:34 /images/animated_gif12_vips_out_master.gif
-rw-r--r-- 1 *** ***   631936 May 27 06:50 /images/animated_gif12_vips_out_new.gif
-rw-r--r-- 1 *** ***  1051173 May 27 06:34 /images/animated_gif13_vips_out_master.gif
-rw-r--r-- 1 *** ***   997631 May 27 06:50 /images/animated_gif13_vips_out_new.gif
-rw-r--r-- 1 *** ***  3355198 May 27 06:51 /images/animated_gif15_vips_out_master.gif
-rw-r--r-- 1 *** ***  3320212 May 27 06:50 /images/animated_gif15_vips_out_new.gif
-rw-r--r-- 1 *** *** 14747250 May 27 06:35 /images/animated_gif16_vips_out_master.gif
-rw-r--r-- 1 *** *** 14557252 May 27 06:51 /images/animated_gif16_vips_out_new.gif
```

### Findings

As we see, the memory footprint is more or less the same for both approaches, yet my approach is slightly faster and produces slightly smaller files. Also, this approach doesn't have a frame size limitation.