@jcupitt 2022-05-27
Oh, nice! It seems to work well on the test images I tried. @dloebl, any thoughts?