@dloebl 2022-05-30
Looks like a nice improvement to me 👍
I've added some comments inline.