@DarthSim 2022-06-04
Could you give a sample, please? I have like 16 test GIFs, and all of them work fine.