@jcupitt 2022-06-04
Sorry, it's not a GIF I can share :( I'm trying to find another example, but no luck so far.

Let's fix this up and merge it, then try to resolve any fails in another PR.