@jcupitt 2022-06-06
I made a GIF I can share that fails.

This terrible GIF:

http://www.rollthepotato.net/~john/skate.gif

If you run this:

```
$ vips copy skate.gif[n=-1] x.gif
```

Becomes:

http://www.rollthepotato.net/~john/x.gif

I think every pixel which is equal to the previous pixel becomes transparent in the output, which causes a lot of artefacts at the start of the clip.