@DarthSim 2022-06-07
Rebased and squashed the branch.

> I made a GIF I can share that fails.

Oh, thank you! I'll dig it.