@DarthSim 2022-06-07
I've tested `vips copy` built from this branch, and I didn't get any artifacts: http://img.darthsim.me/ufHVruZ3o2.gif. I've tested it with both Quantizr and LIQ from the Debian package.

@jcupitt could you test it again on your end? Maybe it was fixed after the rebase?