@jcupitt 2022-06-07
I wiped everything and rebuilt, and I see still bad artefacts.

Could it be the cgif version? I'm using git master.