@dloebl 2022-06-07
Maybe it's similar to the issue mentioned [here](https://github.com/libvips/libvips/issues/2790#issuecomment-1141338058)?
I can reproduce the artefact with libimagequant 2.17 - but only if dithering is enabled. With Quantizr/libimagequant 4.0 the issue is gone on my side.