@jcupitt 2022-06-07
Ah good spot.

But these artefacts only appear with this branch -- master seems to work fine. I think something in this PR has triggered this behaviour.