@DarthSim 2022-06-07
BTW it seems like the same problem may occur when we reuse the global palette from the source image, too.