@jcupitt 2022-06-08
That's great! All my tests pass too, well done.

I see some small leaks in valigrind, but maybe that's just libimagequant. Let's merge and maybe try to patch them in another PR.