@jcupitt 2022-06-08
I did a quick benchmark against imagemagick:

```
$ /usr/bin/time -f %M:%e vipsthumbnail 3198.gif[n=-1] -o vips-13.gif --size 224
57344:0.62
$ /usr/bin/time -f %M:%e convert 3198.gif -resize 75% im.gif
200796:6.44
$ ls -l im.gif vips-13.gif 
-rw-rw-r-- 1 john john 3176859 Jun  8 16:26 im.gif
-rw-r--r-- 1 john john 2487189 Jun  8 16:25 vips-13.gif
```

So for this GIF at least (a short video clip) libvips is now 10x faster, needs 4x less memory, and makes GIFs which are 20% smaller!

Nice work everyone.