@jcupitt 2022-06-08
.... and the dithering of the libvips one is better too, so they look nicer.